+++ fabric/src/main/java/net/craftsupport/anticrasher/fabric/user/FabricUserManager.java
@@ -14,19 +14,19 @@ public class FabricUserManager implements UserManager {
 
     @Override
     public User get(UUID uuid) {
-        if (uuid == null) return new FabricUser(null, UUID.randomUUID(), null);
+        if (uuid == null) return new FabricUser(UUID.randomUUID(), null);
 
         return userCache.get(uuid);
     }
 
     @Override
-    public @NotNull User getOrCreate(com.github.retrooper.packetevents.protocol.player.User user, Object source) {
-        return userCache.containsKey(user.getUUID()) ? get(user.getUUID()) : create(user, source);
+    public @NotNull User getOrCreate(UUID uuid, Object source) {
+        return userCache.containsKey(uuid) ? get(uuid) : create(uuid, source);
     }
 
     @Override
-    public User create(com.github.retrooper.packetevents.protocol.player.User user, Object source) {
-        return userCache.computeIfAbsent(user.getUUID(), id -> new FabricUser(user, id, source));
+    public User create(UUID uuid, Object source) {
+        return userCache.computeIfAbsent(uuid, id -> new FabricUser(id, source));
     }
 
     @Override