+++ .github/CHANGELOG.md
@@ -1,8 +1,9 @@
-# 🚀 AntiCrasher v2.0.2
-A small bug-fix where bStats Metrics would not be initialised for Bukkit and Velocity.
+# 🚀 AntiCrasher v2.0.3
+A small bug-fix, aiming to resolve plugin compat issues with fake players.
 
 ## Changelog
-- Initialised Metrics
+- Fix some plugin incompatibilities
+- Check for NPC or fake-player metadata [Bukkit]
 
 ## Patches
 The following exploits are currently patched in this plugin/mod: